backend/requirements-ci.txt
# CI/CD Testing Requirements
# Excludes heavy dependencies that need special installation (freqtrade, torch)

# FastAPI and server
fastapi==0.121.3
uvicorn[standard]==0.27.1
python-multipart==0.0.22

# Database
asyncpg==0.29.0
supabase==2.3.4
python-dotenv==1.0.1

# Redis (for agent pub/sub)
redis==5.0.1
hiredis==2.3.2

# HTTP client - Fixed version compatibility with supabase
httpx>=0.24.0,<0.26.0
aiohttp>=3.9.0
websockets==12.0

# Crypto exchange libraries
ccxt>=4.5.4

# Data processing
pandas>=2.2.0,<3.0
numpy>=2.0,<3.0
pydantic==2.6.1

# Scheduling
apscheduler==3.10.4

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Logging
structlog==24.1.0

# Security and encryption
cryptography==46.0.5

# System monitoring
psutil==5.9.8

# Advanced analytics and risk management
scipy>=1.13.0
statsmodels>=0.14.2
scikit-learn>=1.5.0
cvxopt==1.3.2
arch==6.3.0
yfinance==0.2.28
ta==0.11.0

# Machine Learning - Lightweight for CI
lightgbm>=4.0.0
xgboost>=2.0.0

# Technical Analysis
# ta-lib and pandas-ta excluded from CI: ta-lib needs system C library,
# pandas-ta 0.4.71b0 requires Python >=3.12 (breaks 3.11 matrix).
# Neither is imported in backend/app/ — used only by FreqTrade strategies.

# Async utilities
aiocache>=0.12.0
aiofiles>=23.0.0

# Rate limiting
slowapi>=0.1.9

# Monitoring and metrics
prometheus-client>=0.19.0

# Distributed tracing (OpenTelemetry)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Error tracking
sentry-sdk[fastapi]>=1.40.0

